refactor(parties/traveller): move traveller to under the parties page and remove old crm from project
