+++ apps/web/src/app/[lang]/app/[type]/navbar/groups.tsx
@@ -320,69 +320,13 @@ export const parties: NavbarItemsFromDB[] = [
     parentNavbarItemKey: "parties",
     displayOrder: 1,
   },
-];
-export const crm: NavbarItemsFromDB[] = [
   {
-    key: "crm",
-    displayName: "Parties",
-    description: "Manage customer relationship management settings.",
-    href: "crm/merchants",
-    icon: "layer",
-    parentNavbarItemKey: "/",
-    displayOrder: 1,
-  },
-  {
-    key: "crm/merchants",
-    displayName: "Merchants",
-    description: "Manage merchant accounts and details.",
-    href: "crm/merchants",
-    icon: "shop",
-    parentNavbarItemKey: "crm",
-    displayOrder: 1,
-  },
-  {
-    key: "crm/refund-points",
-    displayName: "Refund Points",
-    description: "Manage refund points and settings.",
-    href: "crm/refund-points",
-    icon: "refund",
-    parentNavbarItemKey: "crm",
-    displayOrder: 1,
-  },
-  {
-    key: "crm/customs",
-    displayName: "Customs",
-    description: "Manage customs settings and configurations.",
-    href: "crm/customs",
-    icon: "container",
-    parentNavbarItemKey: "crm",
-    displayOrder: 1,
-  },
-  {
-    key: "crm/tax-free",
-    displayName: "Tax Free",
-    description: "Manage tax-free settings and exemptions.",
-    href: "crm/tax-free",
-    icon: "tax",
-    parentNavbarItemKey: "crm",
-    displayOrder: 1,
-  },
-  {
-    key: "crm/tax-offices",
-    displayName: "Tax Offices",
-    description: "Manage tax office details and settings.",
-    href: "crm/tax-offices",
-    icon: "taxOffice",
-    parentNavbarItemKey: "crm",
-    displayOrder: 1,
-  },
-  {
-    key: "crm/traveller",
+    key: "parties/traveller",
     displayName: "Traveller",
     description: "Manage traveller-related settings.",
-    href: "crm/traveller",
+    href: "parties/traveller",
     icon: "plane",
-    parentNavbarItemKey: "crm",
+    parentNavbarItemKey: "parties",
     displayOrder: 1,
   },
 ];
