+++ apps/web/src/app/[lang]/app/[type]/navbar/projects/unirefund.tsx
@@ -1,5 +1,5 @@
 import type { NavbarItemsFromDB } from "@repo/ui/theme/types";
-import { contracts, crm, management, parties, settings } from "../groups";
+import { contracts, management, parties, settings } from "../groups";
 
 export const unirefundNavbarDataFromDB: NavbarItemsFromDB[] = [
   {
@@ -22,7 +22,6 @@ export const unirefundNavbarDataFromDB: NavbarItemsFromDB[] = [
   },
   ...management,
   ...settings,
-  ...crm,
   ...parties,
   ...contracts,
